Current Set Logic
The current set is now accurately displaying. This will eventually need to be changed. I also added logic to remove the "settings box"--which contains the input boxes for the work, rest, and sets--when the user hits start. This way the user won't have to see their config settings while the timer is going.