Resolved zero-time bug
The timer appears to behave appropriately when it reaches zero. I believe the root cause of this bug was due to lack of refreshing of the display when a timer  event reaches zero and the change to the next timer event takes place. I added another line to set the display time to provide an additional refresh before the timer event changes.